pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cascade_guard"
version = "0.4.0"
description = "Cascade Guard environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # OSM data fetching + graph analysis
    "requests>=2.25.0",
    "networkx>=3.0",
    # Pydantic + server
    "pydantic>=2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m cascade_guard.server.app
server = "cascade_guard.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["cascade_guard", "cascade_guard.server"]
package-dir = { "cascade_guard" = ".", "cascade_guard.server" = "server" }